requirements.txt
aiopg==0.16.0
Click==7.0
fastapi==0.33.0
h11==0.8.1
httptools==0.0.13
peewee==2.10.2
peewee-async==0.5.12
psycopg2-binary==2.8.3
pydantic==0.30
starlette==0.12.0
uvicorn==0.8.4
uvloop==0.12.2
websockets==7.0
